--- packages/rs-dpp/src/data_contract/document_type/methods/contested_vote_poll_for_document/mod.rs
@@ -1,1 +1,0 @@
-mod v0;
