--- packages/rs-dpp/src/data_contract/document_type/methods/create_document_from_data/mod.rs
@@ -1,1 +1,0 @@
-mod v0;
