--- packages/rs-dpp/src/data_contract/document_type/methods/create_document_with_prevalidated_properties/mod.rs
@@ -1,1 +1,0 @@
-mod v0;
