--- packages/rs-dpp/src/data_contract/document_type/methods/estimated_size/mod.rs
@@ -1,1 +1,0 @@
-mod v0;
