--- packages/rs-dpp/src/data_contract/document_type/methods/index_for_types/mod.rs
@@ -1,1 +1,0 @@
-mod v0;
