--- packages/rs-dpp/src/data_contract/document_type/methods/max_size/mod.rs
@@ -1,1 +1,0 @@
-mod v0;
