--- packages/rs-dpp/src/data_contract/document_type/methods/serialize_value_for_key/mod.rs
@@ -1,1 +1,0 @@
-mod v0;
