+++ packages/rs-dpp/src/data_contract/document_type/v1/accessors.rs
@@ -1,0 +1,138 @@
+use crate::data_contract::document_type::accessors::{
+    DocumentTypeV0Getters, DocumentTypeV1Getters,
+};
+use crate::data_contract::document_type::index::Index;
+use crate::data_contract::document_type::index_level::IndexLevel;
+use crate::data_contract::document_type::property::DocumentProperty;
+
+use platform_value::{Identifier, Value};
+
+use crate::balances::credits::TokenAmount;
+use crate::data_contract::document_type::restricted_creation::CreationRestrictionMode;
+use crate::data_contract::document_type::token_costs::accessors::TokenCostGettersV0;
+use crate::data_contract::document_type::v1::DocumentTypeV1;
+use crate::data_contract::storage_requirements::keys_for_document_type::StorageKeyRequirements;
+use crate::data_contract::TokenContractPosition;
+use crate::document::transfer::Transferable;
+use crate::identity::SecurityLevel;
+use crate::nft::TradeMode;
+use indexmap::IndexMap;
+use std::collections::{BTreeMap, BTreeSet};
+
+impl DocumentTypeV0Getters for DocumentTypeV1 {
+    fn name(&self) -> &String {
+        &self.name
+    }
+
+    fn schema(&self) -> &Value {
+        &self.schema
+    }
+
+    fn schema_owned(self) -> Value {
+        self.schema
+    }
+
+    fn indexes(&self) -> &BTreeMap<String, Index> {
+        &self.indices
+    }
+
+    fn find_contested_index(&self) -> Option<&Index> {
+        self.indices
+            .iter()
+            .find(|(_, index)| index.contested_index.is_some())
+            .map(|(_, contested_index)| contested_index)
+    }
+
+    fn index_structure(&self) -> &IndexLevel {
+        &self.index_structure
+    }
+
+    fn flattened_properties(&self) -> &IndexMap<String, DocumentProperty> {
+        &self.flattened_properties
+    }
+
+    fn properties(&self) -> &IndexMap<String, DocumentProperty> {
+        &self.properties
+    }
+
+    fn identifier_paths(&self) -> &BTreeSet<String> {
+        &self.identifier_paths
+    }
+
+    fn binary_paths(&self) -> &BTreeSet<String> {
+        &self.binary_paths
+    }
+
+    fn required_fields(&self) -> &BTreeSet<String> {
+        &self.required_fields
+    }
+    fn transient_fields(&self) -> &BTreeSet<String> {
+        &self.transient_fields
+    }
+
+    fn documents_keep_history(&self) -> bool {
+        self.documents_keep_history
+    }
+
+    fn documents_mutable(&self) -> bool {
+        self.documents_mutable
+    }
+
+    fn documents_can_be_deleted(&self) -> bool {
+        self.documents_can_be_deleted
+    }
+
+    fn documents_transferable(&self) -> Transferable {
+        self.documents_transferable
+    }
+
+    fn trade_mode(&self) -> TradeMode {
+        self.trade_mode
+    }
+
+    fn creation_restriction_mode(&self) -> CreationRestrictionMode {
+        self.creation_restriction_mode
+    }
+
+    fn data_contract_id(&self) -> Identifier {
+        self.data_contract_id
+    }
+
+    fn requires_identity_encryption_bounded_key(&self) -> Option<StorageKeyRequirements> {
+        self.requires_identity_encryption_bounded_key
+    }
+
+    fn requires_identity_decryption_bounded_key(&self) -> Option<StorageKeyRequirements> {
+        self.requires_identity_decryption_bounded_key
+    }
+
+    fn security_level_requirement(&self) -> SecurityLevel {
+        self.security_level_requirement
+    }
+}
+
+impl DocumentTypeV1Getters for DocumentTypeV1 {
+    fn document_creation_token_cost(&self) -> Option<(TokenContractPosition, TokenAmount)> {
+        self.token_costs.document_creation_token_cost()
+    }
+
+    fn document_replacement_token_cost(&self) -> Option<(TokenContractPosition, TokenAmount)> {
+        self.token_costs.document_replacement_token_cost()
+    }
+
+    fn document_deletion_token_cost(&self) -> Option<(TokenContractPosition, TokenAmount)> {
+        self.token_costs.document_deletion_token_cost()
+    }
+
+    fn document_transfer_token_cost(&self) -> Option<(TokenContractPosition, TokenAmount)> {
+        self.token_costs.document_transfer_token_cost()
+    }
+
+    fn document_price_update_token_cost(&self) -> Option<(TokenContractPosition, TokenAmount)> {
+        self.token_costs.document_price_update_token_cost()
+    }
+
+    fn document_purchase_token_cost(&self) -> Option<(TokenContractPosition, TokenAmount)> {
+        self.token_costs.document_purchase_token_cost()
+    }
+}
