+++ packages/rs-drive/src/drive/tokens/distribution/add_pre_programmed_distribution/v0/mod.rs
@@ -2,7 +2,6 @@ use crate::drive::tokens::paths::{
     token_ms_timed_at_time_distributions_path_vec, token_ms_timed_distributions_path_vec,
     token_pre_programmed_at_time_distribution_path_vec, token_pre_programmed_distributions_path,
     token_root_pre_programmed_distributions_path,
-    TOKEN_PERPETUAL_DISTRIBUTIONS_FOR_IDENTITIES_LAST_CLAIM_KEY,
     TOKEN_PRE_PROGRAMMED_DISTRIBUTIONS_FOR_IDENTITIES_LAST_CLAIM_KEY,
     TOKEN_PRE_PROGRAMMED_DISTRIBUTIONS_KEY,
 };
